componente skills

+++ ProyectoIntegrador_CabreraCynthia/BACK END/CynthiaMC/src/main/java/com/portfolio/CynthiaMC/Repository/REducacion.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface REducacion extends JpaRepository<Educacion, Integer> {
+public interface REducacion extends JpaRepository<Educacion, Long> {
     public Optional<Educacion> findByNombreE(String nombreE);
     public boolean existsByNombreE(String nombreE);
 }
